@@ -2,6 +2,7 @@
   "private": true,
   "name": "abdurashid.tech",
   "version": "5.0.0",
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -14,8 +15,7 @@
     "framer-motion": "^6.4.3",
     "gray-matter": "^4.0.3",
     "kbar": "^0.1.0-beta.40",
-    "latest": "^0.2.0",
-    "next": "^14.0.4",
+    "next": "^14.2.3",
     "next-seo": "^6.4.0",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
