@@ -1,24 +1,13 @@
 "use client"
 import { styled } from "../../stitches.config";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import { motion, AnimateSharedLayout } from "framer-motion";
 import { useKBar } from "kbar";
 
 export default function Navbar() {
-  const router = useRouter();
-  const [isRouterReady, setIsRouterReady] = useState(false);
-
-  useEffect(() => {
-    if (router.isReady) {
-      setIsRouterReady(true);
-    }
-  }, [router.isReady]);
-
-  if (!isRouterReady) {
-    return null; // or a loading spinner, etc.
-  }
+  const pathname = usePathname();
   const pages = ["About", "Skills", "Projects", "Setup"];
   const [hovered, setHovered] = useState("");
   const { query } = useKBar();
@@ -26,54 +15,54 @@ export default function Navbar() {
   return (
     <AnimateSharedLayout>
       <Header>
-        <Link href="/" passHref>
-          <ButtonLogo as="a">
+        <Link href="/" passHref legacyBehavior>
+          <ButtonLogo>
             <i className="ri-home-2-line" />
           </ButtonLogo>
         </Link>
 
         <Nav>
-          <List>
-            {pages.map((page) => {
-              const path = `/${page.toLowerCase()}`;
-              const isHovered = hovered === page;
+        <List>
+          {pages.map((page) => {
+            const path = `/${page.toLowerCase()}`;
+            const isHovered = hovered === page;
 
-              return (
-                <li key={page}>
-                  <Link href={path} passHref>
-                    <Anchor>
-                      <div
-                        className={`relative inline-block px-5 uppercase transition-colors duration-200 ease-in-out hover:text-primary text-secondary font-medium text-sm tracking-wider cursor-pointer ${
-                          router.pathname == path ? "text-primary" : ""
-                        }`}
-                        onMouseEnter={() => setHovered(page)}
-                        onMouseLeave={() => setHovered("")}
-                      >
-                        {isHovered && (
-                          <motion.span
-                            className="absolute top-[-15px] left-0 right-0 bg-[var(--hover)] p-5 rounded-[var(--borderRadius)] z-[-1]"
-                            layoutId="nav"
-                            initial={{ opacity: 0 }}
-                            animate={{ opacity: 1 }}
-                            exit={{ opacity: 0 }}
-                          />
+            return (
+              <li key={page}>
+                <Link href={path} passHref legacyBehavior>
+                  <Anchor>
+                    <div
+                      className={`relative inline-block px-5 uppercase transition-colors duration-200 ease-in-out hover:text-primary text-secondary font-medium text-sm tracking-wider cursor-pointer ${
+                        pathname === path ? "text-primary" : ""
+                      }`}
+                      onMouseEnter={() => setHovered(page)}
+                      onMouseLeave={() => setHovered("")}
+                    >
+                      {isHovered && (
+                        <motion.span
+                          className="absolute top-[-15px] left-0 right-0 bg-[var(--hover)] p-5 rounded-[var(--borderRadius)] z-[-1]"
+                          layoutId="nav"
+                          initial={{ opacity: 0 }}
+                          animate={{ opacity: 1 }}
+                          exit={{ opacity: 0 }}
+                        />
                       )}
                       {page}
                     </div>
-                    </Anchor>
-                  </Link>
-                </li>
-              );
-            })}
-          </List>
-        </Nav>
+                  </Anchor>
+                </Link>
+              </li>
+            );
+          })}
+        </List>
+      </Nav>
 
         <Aside>
           <ButtonHeader
             as="button"
             type="button"
             aria-label="Command"
-            onClick={query.toggle}
+            onClick={query?.toggle}
             css={{ padding: "0 8px" }}
           >
             <Icon className="ri-command-line" />
@@ -157,44 +146,4 @@ const Anchor = styled("a", {
   border: 0,
   position: "relative",
   "&:hover, &:focus": { opacity: 1 },
-});
-
-const NavContainer = styled(motion.span, {
-  color: "$secondary",
-  cursor: "pointer",
-  display: "inline-block",
-  fontSize: "12px",
-  fontWeight: 500,
-  letterSpacing: "1.2px",
-  padding: "20px",
-  textDecoration: "none",
-  textTransform: "uppercase",
-  transition: "color $duration ease-in-out",
-  "&:hover": {
-    color: "$primary",
-  },
-  "&::after": {
-    content: '""',
-    position: "absolute",
-    margin: "0px auto",
-    top: "18px",
-    left: "0px",
-    right: "0px",
-    height: "1px",
-    width: "20px",
-    background: "rgb(255, 255, 255)",
-    opacity: 0,
-    transition: "opacity $duration ease-in-out",
-  },
-});
-
-const NavHovered = styled(motion.span, {
-  position: "absolute",
-  top: "-15px",
-  left: "0",
-  right: "0",
-  background: "$hover",
-  padding: 20,
-  borderRadius: "$borderRadius",
-  zIndex: -1,
 });
