@@ -1,3 +1,4 @@
+// _app.tsx
 import '../public/static/css/global.css'
 import '../public/static/css/prism.css'
 import 'remixicon/fonts/remixicon.css'
@@ -8,16 +9,14 @@ import CommandBar from '../components/CommandBar'
 
 Router.events.on('routeChangeComplete', url => gtag.pageview(url))
 
-const Noop = ({ children } : {children: any} ) => children
-
-export default function MyApp({ Component, pageProps } : {Component: any, pageProps: any}) {
-  const Layout = typeof Component.Layout === 'function' ? Component.Layout : Noop
-
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
-      <Layout>
+    <html>
+      <body>
         <CommandBar>
-          <Component {...pageProps} />
+          {children}
         </CommandBar>
-      </Layout>
+      </body>
+    </html>
   )
 }
