@@ -1,6 +1,7 @@
 import { parseISO, format } from "date-fns";
+import {ReactElement} from "react";
 
-export default function BlogDate({ dateString }) {
+export default function BlogDate({ dateString }:{dateString: string}): ReactElement {
   if (!dateString) return <div />;
 
   const date = parseISO(dateString);
