@@ -1,4 +1,4 @@
-import { styled } from '../stitches.config'
+import { styled } from '../../stitches.config'
 
 export const FeaturedProjects = styled('div', {
   margin: '10px 0 0 -20px',
