@@ -1,3 +1,4 @@
+"use client"
 import { useState, useEffect } from 'react'
 import { useKBar } from 'kbar'
 import { ButtonPrimary } from '../components/ButtonPrimary'
@@ -16,19 +17,19 @@ export default function ShortcutHome() {
 
     if (isMobile) {
       return (
-        <ButtonPrimary as="button" onClick={query.toggle}>
+        <ButtonPrimary as="button" onClick={query?.toggle}>
           Tap to start →
         </ButtonPrimary>
       )
     } else if (isMac) {
       return (
-        <ButtonPrimary as="button" onClick={query.toggle}>
+        <ButtonPrimary as="button" onClick={query?.toggle}>
           Press <kbd>⌘</kbd> <kbd>K</kbd> to start →
         </ButtonPrimary>
       )
     } else {
       return (
-        <ButtonPrimary as="button" onClick={query.toggle}>
+        <ButtonPrimary as="button" onClick={query?.toggle}>
           Press <kbd>ctrl</kbd> <kbd>K</kbd> to start →
         </ButtonPrimary>
       )
